requirements.txt
#streamlit==1.29.0
#pandas==2.1.3
#plotly==5.18.0
#supabase==2.0.3
#python-dotenv==1.0.0

streamlit==1.28.0
supabase==1.0.3
python-dotenv==1.0.0
gotrue==1.3.1
postgrest==0.10.6
storage3==0.5.3
supafunc==0.2.2
#pip install openpyxl



